requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
dj-database-url==1.3.0
Django==4.1.7
django-appconf==1.0.5
django-auth==0.1.9
gunicorn==20.1.0
Pillow==9.5.0
pkg_resources==0.0.0
psycopg2-binary==2.9.6
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
